% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn.to.igraph.R
\name{bn.to.igraph}
\alias{bn.to.igraph}
\title{Bayesian Network to iGraph Model}
\usage{
bn.to.igraph(bayesian.network)
}
\arguments{
\item{bayesian.network}{Bayesian Network structure.}
}
\description{
Converts Bayesian network structure based on package "bnlearn"
and "bnviewer" to model based on package "igraph".
}
\references{
See online documentation \url{http://robsonfernandes.net/bnviewer}
}
