\name{bnviewer-package}
\alias{bnviewer-package}
\alias{bnviewer}
\docType{package}
\title{Interactive Visualization of Bayesian Networks}
\description{

   Bayesian networks provide an intuitive framework for probabilistic reasoning
             and its graphical nature can be interpreted quite clearly. Graph based methods
             of machine learning are becoming more popular because they offer a richer model
             of knowledge that can be understood by a human in a graphical format. The 'bnviewer'
             is an R Package that allows the interactive visualization of Bayesian Networks.
             The aim of this package is to improve the Bayesian Networks visualization over
             the basic and static views offered by existing packages.

}
\details{

  \tabular{ll}{

    Package: \tab bnviewer\cr
    Type: \tab Package\cr
    Version: \tab 0.1.5\cr
    Date: \tab 2020-07-22\cr
    License: \tab MIT + file LICENSE

  }

}
\author{

  Robson Fernandes\cr
  Institute of Mathematical and Computer Sciences\cr
  University of Sao Paulo - ICMC-USP

  Maintainer: Robson Fernandes \email{robson.fernandes@usp.br}

}
