% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnviewer.R
\name{plot}
\alias{plot}
\title{Interactive Bayesian Network Plots}
\usage{
plot(bayesianNetwork, bayesianNetwork.title = "",
  bayesianNetwork.subtitle = "", bayesianNetwork.footer = "",
  bayesianNetwork.layout = "default", bayesianNetwork.width = "100\%",
  bayesianNetwork.height = "500px", node.shape = c("dot"),
  node.label.prefix = "", node.colors = list(), edges.smooth = TRUE,
  edges.dashes = TRUE, options.highlightNearest = TRUE,
  options.nodesIdSelection = FALSE)
}
\arguments{
\item{bayesianNetwork}{A structure of a Bayesian Network
(Example : hill-climbing (HC)).}

\item{bayesianNetwork.layout}{A layout of a Bayesian Network.
Example :
\enumerate{
  \item layout_on_sphere
  \item layout_on_grid
  \item layout_in_circle
  \item layout_as_star
  \item layout_as_tree
  \item layout_with_sugiyama
  \item layout_with_kk
  \item layout_with_dh
  \item layout_with_lgl
  \item layout_with_mds
  \item layout_with_gem
  \item layout_nicely
  \item layout_components
}}
}
\description{
Interactive Bayesian Network Plots
}
