% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnviewer.R
\name{plot}
\alias{plot}
\title{Plot Bayesian Network}
\usage{
plot(bayesian.network, visNetwork.layout = "default",
  visNetwork.width = "100\%", visNetwork.height = "500px",
  visOptions.highlightNearest = TRUE, visOptions.nodesIdSelection = FALSE,
  node.label.prefix = "", edges.smooth = TRUE, edges.dashes = TRUE)
}
\arguments{
\item{bayesian.network}{A structure of a Bayesian Network
(Example : hill-climbing (HC)).}

\item{visNetwork.layout}{A layout of a Bayesian Network.
Example :
\enumerate{
  \item layout_on_sphere
  \item layout_on_grid
  \item layout_in_circle
  \item layout_as_star
  \item layout_as_tree
  \item layout_with_sugiyama
  \item layout_with_kk
  \item layout_with_dh
  \item layout_with_lgl
  \item layout_with_mds
  \item layout_with_gem
  \item layout_nicely
  \item layout_components
}}
}
\description{
Plot Bayesian Network
}
