% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnviewer.R
\name{plot}
\alias{plot}
\title{Interactive Bayesian Network Plots}
\usage{
plot(bayesianNetwork, bayesianNetwork.title = "",
  bayesianNetwork.subtitle = "", bayesianNetwork.footer = "",
  bayesianNetwork.layout = "default", bayesianNetwork.width = "100\%",
  bayesianNetwork.height = "500px", node.shape = c("dot"),
  node.label.prefix = "", node.colors = list(), edges.smooth = TRUE,
  edges.dashes = TRUE, options.highlightNearest = TRUE,
  options.nodesIdSelection = FALSE)
}
\arguments{
\item{bayesianNetwork}{A structure of a Bayesian Network
(Example : hill-climbing (HC)).}

\item{bayesianNetwork.title}{Bayesian Network title}

\item{bayesianNetwork.subtitle}{Bayesian Network subtitle}

\item{bayesianNetwork.footer}{Bayesian Network footer}

\item{bayesianNetwork.layout}{A layout of a Bayesian Network
\enumerate{
  \item layout_on_sphere
  \item layout_on_grid
  \item layout_in_circle
  \item layout_as_star
  \item layout_as_tree
  \item layout_with_sugiyama
  \item layout_with_kk
  \item layout_with_dh
  \item layout_with_lgl
  \item layout_with_mds
  \item layout_with_gem
  \item layout_nicely
  \item layout_components
}}

\item{bayesianNetwork.width}{Bayesian Network width}

\item{bayesianNetwork.height}{Bayesian Network height}

\item{node.shape}{A node shape of a Bayesian Network
\enumerate{
  \item dot (default)
  \item circle
  \item ellipse
  \item database
  \item diamond
  \item square
  \item triangle
  \item box
  \item star
  \item text
}}

\item{node.label.prefix}{Adds a prefix to the node label}
}
\description{
Interactive Bayesian Network Plots
}
