% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnviewer.R
\name{save}
\alias{save}
\title{Save Bayesian Network into HTML file.}
\usage{
save(bayesianNetwork, file = "bayesianNetwork.html")
}
\arguments{
\item{bayesianNetwork}{A structure of a Bayesian Network
(Example : hill-climbing (HC)).}

\item{file}{Name of file. (Example : bayesianNetwork.html).}
}
\description{
Save Bayesian Network into HTML file.
}
