% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnviewer.R
\name{save}
\alias{save}
\title{Saves the Bayesian network generated by the viewer into HTML file}
\usage{
save(bayesianNetworkViewer, file, outputDir = NULL)
}
\arguments{
\item{bayesianNetworkViewer}{A Bayesian network generated by the viewer.}

\item{file}{File name. (Example : bayesianNetwork.html).}

\item{outputDir}{Output directory name. Directory where Bayesian networks will be saved.
By default they are saved to a temporary directory based on tempdir(),
but it is possible to change the default directory.}
}
\description{
Saves the Bayesian network generated by the viewer into HTML file
}
\examples{

library(bnlearn)
library(bnviewer)

data(coronary)
bn.learn.hc = hc(coronary)

bn.viewer.hc.layout_as_star = viewer(bn.learn.hc,
     bayesianNetwork.width = "100\%",
     bayesianNetwork.height = "80vh",
     bayesianNetwork.layout = "layout_as_star",
     bayesianNetwork.title="Discrete Bayesian Network - Coronary",
     bayesianNetwork.subtitle = "Coronary heart disease data set",
     bayesianNetwork.footer = "Fig. 1 - Layout as star",
     edges.smooth = TRUE,
     node.colors = list(background = "#f4bafd",
                        border = "#2b7ce9",
                        highlight = list(background = "#97c2fc",
                                         border = "#2b7ce9"))
)

#Saves in temp directory
save(bn.viewer.hc.layout_as_star, "bayesianNetwork_layout_as_star.html")

#Saves in specific directory
outputDir = "/home/robsonfernandes/Documentos/bayesianNetworks"
save(bn.viewer.hc.layout_as_star, "bayesianNetwork_layout_as_star.html", outputDir = outputDir)

}
