% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnviewer.R
\name{save}
\alias{save}
\title{Saves the Bayesian Network into HTML file}
\usage{
save(bayesianNetwork, file)
}
\arguments{
\item{bayesianNetwork}{A Bayesian Network structure. (Example : hill-climbing (HC)).}

\item{file}{File name. (Example : bayesianNetwork.html).}
}
\description{
Saves the Bayesian Network into HTML file
}
